Moved settings.go to its own package: configs
