+++ tls.go
@@ -6,9 +6,10 @@ import (
 	"errors"
 	"io/ioutil"
 	"net/http"
+	"github.com/Jumpscale/agentcontroller2/configs"
 )
 
-func configureClientCertificates(httpBinding HTTPBinding, server *http.Server) (err error) {
+func configureClientCertificates(httpBinding configs.HTTPBinding, server *http.Server) (err error) {
 	if httpBinding.ClientCertificateRequired() {
 		server.TLSConfig.ClientAuth = tls.RequireAndVerifyClientCert
 		server.TLSConfig.ClientCAs = x509.NewCertPool()
@@ -26,7 +27,7 @@ func configureClientCertificates(httpBinding HTTPBinding, server *http.Server) (
 	return
 }
 
-func configureServerCertificates(httpBinding HTTPBinding, server *http.Server) (err error) {
+func configureServerCertificates(httpBinding configs.HTTPBinding, server *http.Server) (err error) {
 	server.TLSConfig.Certificates = make([]tls.Certificate, len(httpBinding.TLS), len(httpBinding.TLS))
 	for certificateIndex, tlsSetting := range httpBinding.TLS {
 		certificate, err := tls.LoadX509KeyPair(tlsSetting.Cert, tlsSetting.Key)
